Move quantifiers attributes out of TermDb and into QuantAttribute.
